Test Post Updates on API endpoint
